controllers-test: Move object routines to new file
